refactor: move mutations out of queries.rs
Functions that call a GraphQL mutation should be in a separate file from
queries.rs which as the name implies, is for GraphQL queries.

Fixes #69
