+++ src/graphql/mod.rs
@@ -16,4 +16,5 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 pub mod models;
+pub mod mutations;
 pub mod queries;
